refactor : 리뷰 개선사항 반영 - ErrorMessage 통합 - MenuPrice 생성자 수정 - COMPARE_NUM 제거 - pk seq -> id - MenuProducts 컬렉션 변경 - isCookingOrMeal -> checkCookingOrMeal

+++ src/main/java/kitchenpos/menu/dto/MenuRequest.java
@@ -5,19 +5,18 @@ import kitchenpos.menu.domain.MenuGroup;
 import kitchenpos.menu.domain.MenuProduct;
 import kitchenpos.menu.domain.Product;
 
-import java.math.BigDecimal;
 import java.util.List;
 import java.util.stream.Collectors;
 
 public class MenuRequest {
     private String name;
-    private BigDecimal price;
+    private Integer price;
     private Long menuGroupId;
     private List<MenuProductRequest> menuProductRequests;
 
     public MenuRequest() {}
 
-    public MenuRequest(String name, BigDecimal price, Long menuGroupId, List<MenuProductRequest> menuProductRequests) {
+    public MenuRequest(String name, Integer price, Long menuGroupId, List<MenuProductRequest> menuProductRequests) {
         this.name = name;
         this.price = price;
         this.menuGroupId = menuGroupId;
@@ -37,7 +36,7 @@ public class MenuRequest {
         return name;
     }
 
-    public BigDecimal getPrice() {
+    public Integer getPrice() {
         return price;
     }
